pdf_fill/requirement.txt
arabic-reshaper==3.0.0
et-xmlfile==1.1.0
openpyxl==3.1.2
Pillow==9.5.0
pypdf2==3.0.1
python-bidi==0.4.2
reportlab==3.6.12
six==1.16.0
typing-extensions==4.5.0
